// src/lib/api/generated/models/packItemInputDtoSelectionMode.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Beauty Pack Recommendation API
 * Backend API for beauty-pack personalization, recommendations, COD orders, and administration
 * OpenAPI spec version: 1.0.0
 */

export type PackItemInputDtoSelectionMode = typeof PackItemInputDtoSelectionMode[keyof typeof PackItemInputDtoSelectionMode];


export const PackItemInputDtoSelectionMode = {
  FIXED_REFERENCE: 'FIXED_REFERENCE',
  AUTO_BEST_REFERENCE: 'AUTO_BEST_REFERENCE',
  CUSTOMER_CHOICE: 'CUSTOMER_CHOICE',
} as const;